+++ src/main/java/ca/corefacility/bioinformatics/irida/ria/security/CredentialsExpriredAuthenticationFailureHandler.java
@@ -19,7 +19,7 @@ import org.springframework.stereotype.Component;
 import ca.corefacility.bioinformatics.irida.model.user.PasswordReset;
 import ca.corefacility.bioinformatics.irida.model.user.Role;
 import ca.corefacility.bioinformatics.irida.model.user.User;
-import ca.corefacility.bioinformatics.irida.ria.web.PasswordResetController;
+import ca.corefacility.bioinformatics.irida.ria.web.login.PasswordResetController;
 import ca.corefacility.bioinformatics.irida.service.user.PasswordResetService;
 import ca.corefacility.bioinformatics.irida.service.user.UserService;