+++ src/main/java/ca/corefacility/bioinformatics/irida/ria/web/users/UsersAjaxController.java
@@ -8,11 +8,16 @@ import javax.servlet.http.HttpServletRequest;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
 
+import ca.corefacility.bioinformatics.irida.model.user.PasswordReset;
 import ca.corefacility.bioinformatics.irida.model.user.User;
+import ca.corefacility.bioinformatics.irida.ria.web.ajax.dto.ajax.AjaxErrorResponse;
 import ca.corefacility.bioinformatics.irida.ria.web.ajax.dto.ajax.AjaxFormErrorResponse;
 import ca.corefacility.bioinformatics.irida.ria.web.ajax.dto.ajax.AjaxResponse;
+import ca.corefacility.bioinformatics.irida.ria.web.ajax.dto.ajax.AjaxSuccessResponse;
+import ca.corefacility.bioinformatics.irida.ria.web.exceptions.UIEmailSendException;
 import ca.corefacility.bioinformatics.irida.ria.web.models.tables.TableResponse;
 import ca.corefacility.bioinformatics.irida.ria.web.services.UIUsersService;
 import ca.corefacility.bioinformatics.irida.ria.web.users.dto.*;
@@ -23,11 +28,11 @@ import ca.corefacility.bioinformatics.irida.ria.web.users.dto.*;
 @RestController
 @RequestMapping("/ajax/users")
 public class UsersAjaxController {
-	private final UIUsersService UIUsersService;
+	private final UIUsersService uiUsersService;
 
 	@Autowired
-	public UsersAjaxController(UIUsersService UIUsersService) {
-		this.UIUsersService = UIUsersService;
+	public UsersAjaxController(UIUsersService uiUsersService) {
+		this.uiUsersService = uiUsersService;
 	}
 
 	/**
@@ -38,7 +43,7 @@ public class UsersAjaxController {
 	 */
 	@RequestMapping("/list")
 	public TableResponse<UserDetailsModel> getUsersPagedList(@RequestBody AdminUsersTableRequest request) {
-		return UIUsersService.getUsersPagedList(request);
+		return uiUsersService.getUsersPagedList(request);
 	}
 
 	/**
@@ -72,7 +77,7 @@ public class UsersAjaxController {
 	@RequestMapping("/edit")
 	public ResponseEntity<String> updateUserStatus(@RequestParam Long id, @RequestParam boolean isEnabled,
 			Locale locale) {
-		return UIUsersService.updateUserStatus(id, isEnabled, locale);
+		return uiUsersService.updateUserStatus(id, isEnabled, locale);
 	}
 
 	/**
@@ -88,7 +93,7 @@ public class UsersAjaxController {
 	public ResponseEntity<AjaxResponse> updateUser(@PathVariable Long userId,
 			@RequestBody UserEditRequest userEditRequest, Principal principal, HttpServletRequest request) {
 
-		UserDetailsResponse response = UIUsersService.updateUser(userId, userEditRequest, principal, request);
+		UserDetailsResponse response = uiUsersService.updateUser(userId, userEditRequest, principal, request);
 
 		if (response.hasErrors())
 			return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(new AjaxFormErrorResponse(response.getErrors()));
@@ -111,7 +116,7 @@ public class UsersAjaxController {
 	public ResponseEntity<AjaxResponse> changeUserPassword(@PathVariable Long userId, @RequestParam String oldPassword,
 			@RequestParam String newPassword, Principal principal, HttpServletRequest request) {
 
-		UserDetailsResponse response = UIUsersService.changeUserPassword(userId, oldPassword, newPassword, principal,
+		UserDetailsResponse response = uiUsersService.changeUserPassword(userId, oldPassword, newPassword, principal,
 				request);
 
 		if (response.hasErrors())
@@ -132,6 +137,26 @@ public class UsersAjaxController {
 	public ResponseEntity<UserDetailsResponse> getUserDetails(@PathVariable("userId") Long userId,
 			@RequestParam(value = "mailFailure", required = false, defaultValue = "false") final Boolean mailFailure,
 			Principal principal) {
-		return ResponseEntity.ok(UIUsersService.getUser(userId, mailFailure, principal));
+		return ResponseEntity.ok(uiUsersService.getUser(userId, mailFailure, principal));
+	}
+
+	/**
+	 * Create a new {@link PasswordReset} for the given {@link User}
+	 *
+	 * @param userId    The ID of the {@link User}
+	 * @param principal a reference to the logged in user.
+	 * @param locale    a reference to the locale specified by the browser.
+	 * @return text to display to the user about the result of creating a password reset.
+	 */
+	@RequestMapping(value = "/{userId}/reset-password", method = RequestMethod.POST)
+	@PreAuthorize("hasAnyRole('ROLE_ADMIN', 'ROLE_MANAGER')")
+	public ResponseEntity<AjaxResponse> adminNewPasswordReset(@PathVariable Long userId, Principal principal,
+			Locale locale) {
+		try {
+			return ResponseEntity.ok(
+					new AjaxSuccessResponse(uiUsersService.adminNewPasswordReset(userId, principal, locale)));
+		} catch (UIEmailSendException e) {
+			return ResponseEntity.status(HttpStatus.CONFLICT).body(new AjaxErrorResponse(e.getMessage()));
+		}
 	}
 }