+++ src/main/webapp/resources/js/pages/admin/components/AdminSequencingRunsPage.jsx
@@ -10,17 +10,18 @@
 import React from "react";
 import { PageWrapper } from "../../../components/page/PageWrapper";
 import { PagedTableProvider } from "../../../components/ant.design/PagedTable";
-import { SequencingRunsList } from "../../sequencing-runs/SequencingRunsList";
 import { setBaseUrl } from "../../../utilities/url-utilities";
+import SequencingRunListTable
+  from "../../sequencing-runs/components/SequencingRunListTable";
 
 export default function AdminSequencingRunsPage({}) {
-  const URL = setBaseUrl(`ajax/sequencingRuns`);
+  const URL = setBaseUrl(`ajax/sequencing-runs`);
 
   // The following renders the Sequencing Runs component view
   return (
     <PageWrapper title={i18n("AdminPanel.sequencingRuns")}>
       <PagedTableProvider url={`${URL}/list`}>
-        <SequencingRunsList/>
+        <SequencingRunListTable/>
       </PagedTableProvider>
     </PageWrapper>
   );