+++ src/test/java/ca/corefacility/bioinformatics/irida/web/controller/test/unit/users/UsersAjaxControllerTest.java
@@ -1,6 +1,7 @@
 package ca.corefacility.bioinformatics.irida.web.controller.test.unit.users;
 
 import java.security.Principal;
+import java.util.Locale;
 
 import javax.servlet.http.HttpServletRequest;
 
@@ -12,6 +13,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 
+import ca.corefacility.bioinformatics.irida.model.user.Role;
 import ca.corefacility.bioinformatics.irida.model.user.User;
 import ca.corefacility.bioinformatics.irida.ria.web.ajax.dto.ajax.AjaxResponse;
 import ca.corefacility.bioinformatics.irida.ria.web.services.UIUsersService;
@@ -27,13 +29,13 @@ import static org.mockito.Mockito.*;
 
 public class UsersAjaxControllerTest {
 	private UserService userService;
-	private PasswordResetService passwordResetService;
 	private EmailController emailController;
 	private MessageSource messageSource;
 	private PasswordEncoder passwordEncoder;
 	private HttpServletRequest request;
 	private UIUsersService uiUsersService;
 	private UsersAjaxController controller;
+	private PasswordResetService passwordResetService;
 
 	private final User USER1 = new User(1L, "Elsa", "elsa@arendelle.ca", "Password1!", "Elsa", "Oldenburg", "1234");
 	private final User USER2 = new User(2L, "Anna", "anna@arendelle.ca", "Password2!", "Anna", "Oldenburg", "5678");
@@ -45,9 +47,12 @@ public class UsersAjaxControllerTest {
 		emailController = mock(EmailController.class);
 		messageSource = mock(MessageSource.class);
 		passwordEncoder = new BCryptPasswordEncoder();
+		passwordResetService = mock(PasswordResetService.class);
 		request = mock(HttpServletRequest.class);
-		uiUsersService = new UIUsersService(userService, passwordResetService, emailController, messageSource,
-				passwordEncoder);
+
+		uiUsersService = new UIUsersService(userService, emailController, messageSource, passwordEncoder,
+				passwordResetService);
+
 		controller = new UsersAjaxController(uiUsersService);
 
 		when(userService.read(anyLong())).thenReturn(USER2);
@@ -85,6 +90,22 @@ public class UsersAjaxControllerTest {
 	void getUserDetailsTest() {
 		Principal principal = () -> USER1.getFirstName();
 		ResponseEntity<UserDetailsResponse> response = controller.getUserDetails(USER2.getId(), false, principal);
+		assertEquals(response.getStatusCode(), HttpStatus.OK, "Received an 200 OK response");
+	}
+
+	@Test
+	void adminNewPasswordResetTest() {
+		User user1 = new User(1L, "Elsa", "elsa@arendelle.ca", "Password1!", "Elsa", "Oldenburg", "1234");
+		user1.setSystemRole(Role.ROLE_ADMIN);
+		User user2 = new User(2L, "Anna", "anna@arendelle.ca", "Password2!", "Anna", "Oldenburg", "5678");
+		Principal principal = () -> user1.getFirstName();
+		Locale locale = new Locale("en");
+
+		when(userService.read(anyLong())).thenReturn(user2);
+		when(userService.getUserByUsername(anyString())).thenReturn(user1);
+		when(messageSource.getMessage(anyString(), any(), any(Locale.class))).thenReturn("Anything can work here");
+
+		ResponseEntity<AjaxResponse> response = controller.adminNewPasswordReset(user2.getId(), principal, locale);
 		assertEquals(response.getStatusCode(), HttpStatus.OK, "Received an 200 OK response");
 	}
 }